Rename bsp::uart as apb_uart
