+++ examples/headsail-bsp/Justfile
@@ -1,2 +1,2 @@
 run example target='riscv64imac':
-	cargo run --example {{example}} -Fhpc-rt -Fpanic-uart --target {{target}}-unknown-none-elf
+	cargo run --example {{example}} -Fhpc-rt -Fpanic-apb-uart0 --target {{target}}-unknown-none-elf
