+++ examples/headsail-bsp/examples/uart0-read.rs
@@ -2,11 +2,11 @@
 #![no_main]
 
 extern crate alloc;
-use headsail_bsp::{init_alloc, rt::entry, sprint, sprintln, uart::uart_read_to_heap};
+use headsail_bsp::{apb_uart::uart_read_to_heap, init_alloc, rt::entry, sprint, sprintln};
 
 #[entry]
 fn main() -> ! {
-    sprintln!("Connect to uart with: screen /tmp/uart0");
+    sprintln!("Connect to APB UART 0 with: screen /tmp/uart0");
     init_alloc();
     loop {
         let res = uart_read_to_heap(8);
