+++ examples/headsail-bsp/examples/uart0.rs
@@ -1,7 +1,7 @@
 #![no_std]
 #![no_main]
 
-use headsail_bsp::{rt::entry, uart::uart_write};
+use headsail_bsp::{apb_uart::uart_write, rt::entry};
 
 #[entry]
 fn main() -> ! {
